grpc-client/src/test/python/requirements.txt
gensim>=4.3.0
grpcio>=1.62.1
grpcio-tools>=1.62.1
